[FEAT] #4 - Insert payment
