+++ presentation/build.gradle
@@ -72,4 +72,6 @@ dependencies {
     kapt 'com.google.dagger:hilt-compiler:2.42'
 
     implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.4")
+
+    implementation 'androidx.activity:activity-ktx:1.5.0'
 }
